Made changes to allow the compiler to compile without warnings.

+++ cores/esp8266/cbuf.h
@@ -70,7 +70,7 @@ class cbuf {
             size_t bytes_available = getSize();
             size_t size_to_read = (size < bytes_available) ? size : bytes_available;
             size_t size_read = size_to_read;
-            if(_end < _begin && size_to_read > _bufend - _begin) {
+            if(_end < _begin && size_to_read > (size_t)(_bufend - _begin)) {
                 size_t top_size = _bufend - _begin;
                 memcpy(dst, _begin, top_size);
                 _begin = _buf;
@@ -95,7 +95,7 @@ class cbuf {
             size_t bytes_available = room();
             size_t size_to_write = (size < bytes_available) ? size : bytes_available;
             size_t size_written = size_to_write;
-            if(_end > _begin && size_to_write > _bufend - _end) {
+            if(_end > _begin && size_to_write > (size_t)(_bufend - _end)) {
                 size_t top_size = _bufend - _end;
                 memcpy(_end, src, top_size);
                 _end = _buf;
